domain-search-app/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
bcrypt==4.1.2
Werkzeug==3.0.1
psycopg2-binary==2.9.9
